MAF entry in class locus changed from double -> map<string, double>
This is a first step as to implement MAF estimated separately from cases and
controls (instead of pooled estimate as being used now), provided a case-control
sample is analyzed (todo.txt has been updated with this point). For now, the
column "maf" in the output files has changed to "MAF_pooled".